Remove `Pizzeria.Store.Api`
