+++ sample-tests/Pizzeria.Tests.Integration/Pizzeria.Tests.Integration.csproj
@@ -10,7 +10,6 @@
   <ItemGroup>
     <ProjectReference Include="..\..\sample\Pizzeria.AppHost\Pizzeria.AppHost.csproj" />
     <ProjectReference Include="..\..\sample\Pizzeria.Auth\Pizzeria.Auth.csproj" />
-    <ProjectReference Include="..\..\sample\Pizzeria.Store.Api\Pizzeria.Store.Api.csproj" />
     <ProjectReference Include="..\..\sample\Pizzeria.Store.Data\Pizzeria.Store.Data.csproj" />
     <ProjectReference Include="..\..\sample\Pizzeria.Store.Domain\Pizzeria.Store.Domain.csproj" />
     <ProjectReference Include="..\..\sample\Pizzeria.Store.Web\Pizzeria.Store.Web.csproj" />
