+++ sample/Pizzeria.AppHost/Pizzeria.AppHost.csproj
@@ -12,13 +12,11 @@
     <PackageReference Include="Aspire.Hosting.AppHost" />
     <PackageReference Include="Aspire.Hosting.Keycloak" />
     <PackageReference Include="Aspire.Hosting.PostgreSQL" />
-    <PackageReference Include="Aspire.Hosting.Azure.SignalR" />
   </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\Pizzeria.Common\Pizzeria.Common.csproj" IsAspireProjectResource="false" />
     <ProjectReference Include="..\Pizzeria.Configuration\Pizzeria.Configuration.csproj" />
-    <ProjectReference Include="..\Pizzeria.Store.Api\Pizzeria.Store.Api.csproj" />
     <ProjectReference Include="..\Pizzeria.Store.Web\Pizzeria.Store.Web.csproj" />
   </ItemGroup>
