--- sample/Pizzeria.Store.Api/Pizzeria.Store.Api.csproj
@@ -1,24 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <ItemGroup>
-    <PackageReference Include="Aspire.Keycloak.Authentication" />
-    <PackageReference Include="FastEndpoints" />
-    <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Tools">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-    </PackageReference>
-    <PackageReference Include="Microsoft.Extensions.Diagnostics.HealthChecks.EntityFrameworkCore" />
-    <PackageReference Include="Swashbuckle.AspNetCore" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\src\Lewee.Infrastructure.AspNet.WebApi\Lewee.Infrastructure.AspNet.WebApi.csproj" />
-    <ProjectReference Include="..\..\src\Lewee.Infrastructure.PostgreSQL\Lewee.Infrastructure.PostgreSQL.csproj" />
-    <ProjectReference Include="..\Pizzeria.Common\Pizzeria.Common.csproj" />
-    <ProjectReference Include="..\Pizzeria.Store.Application\Pizzeria.Store.Application.csproj" />
-    <ProjectReference Include="..\Pizzeria.Store.Data\Pizzeria.Store.Data.csproj" />
-    <ProjectReference Include="..\Pizzeria.ServiceDefaults\Pizzeria.ServiceDefaults.csproj" />
-  </ItemGroup>
-
-</Project>
